stt/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Yaron McNabb <y.mcnabb@htm.nl>", "Tung T. Phan <t.phan@htm.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
pydantic = "^2.4.2"
openai = "^1.1.1"
pandas = "^2.1.3"
streamlit = "^1.28.1"
pydub = "^0.25.1"
ffmpeg = "^1.4"
langchain-community = "^0.0.29"
langchain-openai = "^0.1.1"
langchain-core = "^0.1.33"
jiwer = "^3.0.3"
loguru = "^0.7.2"
fastapi = "^0.110.2"
azure-storage-blob = "^12.19.1"
azure-identity = "^1.16.0"
python-multipart = "^0.0.9"
pytest = "^8.2.0"
testcontainers = "^4.4.0"
wiremock = "^2.6.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.7"
pytest-mock = "^3.14.0"
factory-boy = "^3.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
deptry = "^0.16.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
